+++ api/src/helpers/google/README.md
@@ -7,3 +7,7 @@ JWT クライアントをサービスアカウントで認証する
 ### insertEvent.ts
 
 イベントをカレンダーに追加する
+
+### listEvent.ts
+
+イベントのリストを表示する
