+++ api/src/helpers/google/types/listEvent.type.ts
@@ -1,0 +1,4 @@
+export type Events = {
+  date: string | undefined;
+  time: string | undefined;
+}[];
